Handle blocking rejected tasks more safely
Notably, if the work queue is cleared before shutting down, it's
possible that a late item could get to execute out-of-order.

As for the handler itself, if the thread pool is shutdown on a thread
separate from the rejected execution handler, it could be queued after
a shutdown, unless the handler checks whether the pool is shutdown
between the work queue having space and queuing the task.